+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adding anotation XmlTranisent to flagFilterer in no.difi.vefa.validator.api.Section. [#14](https://github.com/difi/vefa-validator/issues/14)
 * XsltChecker changed name to ScrlXsltChecker.
+* Changed signature of no.difi.vefa.validator.api.Source.createInstance(...).
 
 ## 2.0.0